UI/dashboard/requirements.txt
shiny
matplotlib
psycopg2
pandas
python-dotenv
functools
seaborn